Fix bug in RML/MERGE2.  Add comments.
RML/MERGE2 was not handling lists with an odd number of elements. As a
result, It was calling CAR on numbers.  This patch adds two checks for
the cases when either LEFT or RIGHT are just numbers (we already had a
check for the case when LEFT and RIGHT were both numbers).